+++ src/store/configureStore.js
@@ -29,7 +29,7 @@ export default function configureStore(history, initialState) {
   if (__DEVELOPMENT__ && module.hot) {
     // Enable Webpack hot module replacement for reducers
     module.hot.accept('../reducers', () => {
-      const nextRootReducer = require('../reducers')
+      const nextRootReducer = require('../reducers').default
       store.replaceReducer(nextRootReducer)
     })
   }