+++ src/components/Layout.js
@@ -3,8 +3,8 @@ import React, { PureComponent, PropTypes } from 'react'
 import classNames from 'classnames'
 import styles from './Layout.scss'
 import { DARK } from '../constants/index'
-import { Header } from 'twreporter-react-components'
-import { Footer } from 'twreporter-react-components'
+import { Header } from '@twreporter/react-components'
+import { Footer } from '@twreporter/react-components'
 
 // lodash
 import get from 'lodash/get'