+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+- Fix inconsistent line-heights of paragraphs containing annotations
 
 ### 2.2.0
 - Code splitting and dynamic bundle loading by webpack