+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 - Fix Google structure data errors
+- Fix Webpack hot loadig problem on dev
+- Transform import to require.ensure by babel plugin
+- Remove react-hot-loader package since we don't use it at all
 
 ### 2.2.1
 - Update the schema of homepage [#665](https://github.com/twreporter/twreporter-react/pull/665)